apply header revision to fair udp

+++ scratch/udp_application/fair-udp.h
@@ -12,6 +12,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * Author: Chang-Hui Kim <kch9001@gmail.com>
  */
 
 #ifndef UDP_APP_H
@@ -43,7 +45,7 @@ namespace ns3
 
     Ptr<Socket> socket_;
     port_t port_;
-    uint32_t seq_number_{0};
+    uint16_t seq_number_{0};
   };
   
 } // namespace ns3
